refactor: add post as a component
