+++ src/components/common/index.ts
@@ -1,1 +1,2 @@
 export * from './Navigation';
+export * from './Post';
